+++ tokio-reactor/src/async_await.rs
@@ -1,5 +1,0 @@
-use Registration;
-
-use std::marker::Unpin;
-
-impl Unpin for Registration {}